move redirects to their own file in src/redirects.js, add white space between tests
