+++ src/redirects.test.js
@@ -1,5 +1,5 @@
 const url = require('url')
-const { processRedirectString, getRedirect } = require('./server')
+const { processRedirectString, getRedirect } = require('./redirects')
 
 describe('processRedirectString', () => {
   it('reads the regex, replacement and code', () => {
@@ -10,11 +10,13 @@ describe('processRedirectString', () => {
     expect(replace).toEqual('/bar')
     expect(code).toEqual(418)
   })
+
   it('code defaults to 301', () => {
     const { code } = processRedirectString('^/x /x')
 
     expect(code).toEqual(301)
   })
+
   it('detects whether we are matching a pathname or the whole url', () => {
     const { matchPathname } = processRedirectString('^/pathname /x')
     expect(matchPathname).toEqual(true)
@@ -38,6 +40,7 @@ describe('getRedirect', () => {
       getRedirect('www.example.com', '/not-used', { req: fakeReq, dev: false })
     ).toEqual([301, 'https://example.com/foo/bar?baz'])
   })
+
   const itRedirects = (source, target, code = 301) => {
     const addHost = pathOrUrl => {
       if (pathOrUrl.startsWith('/')) {
@@ -71,16 +74,19 @@ describe('getRedirect', () => {
       'https://dvc.org/doc/command-reference/',
       303
     )
+
     itRedirects(
       'https://man.dvc.org/foo',
       'https://dvc.org/doc/command-reference/foo',
       303
     )
+
     itRedirects(
       'https://error.dvc.org/',
       'https://dvc.org/user-guide/troubleshooting#',
       303
     )
+
     itRedirects(
       'https://error.dvc.org/foo',
       'https://dvc.org/user-guide/troubleshooting#foo',
@@ -94,21 +100,25 @@ describe('getRedirect', () => {
       'https://s3-us-east-2.amazonaws.com/dvc-public/code/foo/bar',
       303
     )
+
     itRedirects(
       'https://data.dvc.org/foo/bar',
       'https://s3-us-east-2.amazonaws.com/dvc-public/data/foo/bar',
       303
     )
+
     itRedirects(
       'https://remote.dvc.org/foo/bar',
       'https://s3-us-east-2.amazonaws.com/dvc-public/remote/foo/bar',
       303
     )
+
     itRedirects(
       '/deb/foo',
       'https://s3-us-east-2.amazonaws.com/dvc-s3-repo/deb/foo',
       303
     )
+
     itRedirects(
       '/rpm/foo',
       'https://s3-us-east-2.amazonaws.com/dvc-s3-repo/rpm/foo',
@@ -118,15 +128,21 @@ describe('getRedirect', () => {
 
   describe('discord', () => {
     itRedirects('/help', 'https://discordapp.com/invite/dvwXA2N', 303)
+
     itRedirects('/chat', 'https://discordapp.com/invite/dvwXA2N', 303)
   })
 
   describe('in-site moves', () => {
     itRedirects('/docs/x', '/doc/x')
+
     itRedirects('/documentation/x', '/doc/x')
+
     itRedirects('/doc/commands-reference/add', '/doc/command-reference/add')
+
     itRedirects('/doc/tutorial/', '/doc/tutorials')
+
     itRedirects('/doc/tutorial/subject', '/doc/tutorials/deep/subject')
+
     itRedirects(
       '/doc/use-cases/data-and-model-files-versioning',
       '/doc/use-cases/versioning-data-and-model-files'
